LayoutTests/http/tests/webcodecs/vp9-hdr.html

<!DOCTYPE html>
<html>
<head>
    <title>WebCodecs VP9 HDR decoding</title>
    <script src="../../resources/testharness.js"></script>
    <script src="../../resources/testharnessreport.js"></script>
</head>
<body>
<script>
const HDR_CODEC = "vp09.02.10.10.01.09.16.09.00";
promise_test(async () => {
    assert_implements(window.VideoDecoder, "VideoDecoder is supported");
    const config = { codec: HDR_CODEC, codedWidth: 320, codedHeight: 240 };
    const support = await VideoDecoder.isConfigSupported(config);
    assert_true(support.supported, "VP9 profile 2 (10-bit) is supported");
}, "VP9 profile 2 isConfigSupported");

promise_test(async () => {
    const frames = [];
    const decoder = new VideoDecoder({
        output: frame => frames.push(frame),
        error: e => assert_unreached(`decode error: ${e.message}`),
    });
    decoder.configure({
        codec: HDR_CODEC,
        codedWidth: 320,
        codedHeight: 240,
    });

    // Content generated by ffmpeg, we need to strip the IVF header (32 bytes) and frame header (12 bytes).
    const response = await fetch("vp9-hdr.bin");
    const buffer = await response.arrayBuffer();
    const vp9Data = new Uint8Array(buffer).slice(44);
    decoder.decode(new EncodedVideoChunk({
        type: "key",
        timestamp: 0,
        data: vp9Data,
    }));
    await decoder.flush();

    assert_equals(frames.length, 1, "one frame decoded");
    const frame = frames[0];

    assert_equals(frame.colorSpace.primaries, "bt2020", "primaries");
    assert_equals(frame.colorSpace.matrix, "bt2020-ncl", "matrix");
    assert_false(frame.colorSpace.fullRange, "video range");

    if (frame.format !== "I420P10") {
        assert_equals(frame.format, null);
        assert_true(!window.internals?.is10bitsVideoFrame || internals.is10bitsVideoFrame(frame), "10 bits");
    }

    frame.close();
}, "VP9 profile 2 decoded with WebCodecs has BT.2020 and is 10 bits");
</script>
</body>
</html>
